# openapi.yaml
openapi: 3.0.0
info:
  title: Tours
  version: 0.0.1
servers:
  - url: http://localhost:8080/api/v1/
    description: Dev server
paths:
  /tourist:
    post:
      summary: Создание туриста
      tags:
        - Tourist
      operationId: postTourist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TouristDTO'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TouristDTO"
        'default':
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Изменить данные туриста
      tags:
        - Tourist
      operationId: putHotel
      parameters:
        - name: id
          in: query
          required: true
          description: Индефиктор отеля
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TouristDTO'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotels"
        'default':
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Hotel:
      type: object
      required:
        - name
        - countRooms
        - city
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Sunrise
        countRooms:
          type: integer
          example: 25
        city:
          type: string
          example: Moscow
    Hotels:
      type: array
      items:
        $ref: "#/components/schemas/Hotel"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
...
    TouristDTO:
      type: object
      required:
        - firstName
        - secondName
        - gender
        - age
        - luggage
        - countChildren
      properties:
        id:
          type: integer
          example: 1
        firstName:
          type: string
          example: Oleg
        secondName:
          type: string
          example: Petrov
        gender:
          type: string
          example: male
        age:
          type: integer
          example: 47
        luggage:
          type: boolean
          example: true
        countChildren:
          type: integer
          example: 2
        tourTimesId:
          type: array
          items:
            type: integer
        carRentsId:
          type: array
          items:
            type: integer
        hotelAccommodationsId:
          type: array
          items:
            type: integer
...