# src/main/java/com/bd/tours/dto/TouristDTO.java
package com.bd.tours.dto;

import lombok.*;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@ToString
public class TouristDTO {
    private Long id;
    private String firstName;

    private String secondName;

}